RHOAIENG-72362: Extract genRandomChars to the new @odh-dashboard/foundation
