+++ frontend/src/api/k8s/configMaps.ts
@@ -5,10 +5,10 @@ import {
   k8sUpdateResource,
   K8sStatus,
 } from '@openshift/dynamic-plugin-sdk-utils';
+import { genRandomChars } from '@odh-dashboard/foundation';
 import { KnownLabels } from '@odh-dashboard/k8s-core';
 import { ConfigMapKind, K8sAPIOptions } from '#~/k8sTypes';
 import { ConfigMapModel } from '#~/api/models';
-import { genRandomChars } from '#~/utilities/string';
 import { applyK8sAPIOptions } from '#~/api/apiMergeUtils';
 
 export const CONFIGMAP_PREFIX = 'configmap-';
