+++ frontend/src/api/k8s/roleBindings.ts
@@ -8,6 +8,7 @@ import {
   K8sResourceCommon,
   k8sListResourceItems,
 } from '@openshift/dynamic-plugin-sdk-utils';
+import { genRandomChars } from '@odh-dashboard/foundation';
 import { KnownLabels } from '@odh-dashboard/k8s-core';
 import {
   K8sAPIOptions,
@@ -16,7 +17,6 @@ import {
   RoleBindingSubject,
 } from '#~/k8sTypes';
 import { RoleBindingModel } from '#~/api/models';
-import { genRandomChars } from '#~/utilities/string';
 import { applyK8sAPIOptions } from '#~/api/apiMergeUtils';
 import { RoleBindingPermissionsRoleType } from '#~/concepts/roleBinding/types';
 import { addOwnerReference } from '#~/api/k8sUtils';
