+++ frontend/src/api/k8s/secrets.ts
@@ -8,10 +8,10 @@ import {
   k8sPatchResource,
   K8sResourceCommon,
 } from '@openshift/dynamic-plugin-sdk-utils';
+import { genRandomChars } from '@odh-dashboard/foundation';
 import {
   KnownLabels,
   DATA_CONNECTION_PREFIX,
-  genRandomChars,
   getGeneratedSecretName,
   translateDisplayNameForK8s,
 } from '@odh-dashboard/k8s-core';
