+++ frontend/src/package.json
@@ -37,6 +37,7 @@
     ]
   },
   "dependencies": {
+    "@odh-dashboard/foundation": "*",
     "@odh-dashboard/k8s-core": "*",
     "@odh-dashboard/plugin-core": "*",
     "@odh-dashboard/ui-core": "*"
