+++ frontend/src/pages/pipelines/global/modelCustomization/utils.ts
@@ -4,6 +4,7 @@ import type {
   K8sNameDescriptionFieldData,
   SecretKind,
 } from '@odh-dashboard/k8s-core';
+import { genRandomChars } from '@odh-dashboard/foundation';
 import { getResourceNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import { assembleSecretJudge, assembleSecretTeacher, createSecret } from '#~/api';
 import {
@@ -15,7 +16,6 @@ import {
   ModelCustomizationFormData,
   TeacherJudgeFormData,
 } from '#~/concepts/pipelines/content/modelCustomizationForm/modelCustomizationFormSchema/validationUtils';
-import { genRandomChars } from '#~/utilities/string';
 import { getInputDefinitionParams } from '#~/concepts/pipelines/content/createRun/utils';
 import {
   PipelineVersionKF,
