+++ packages/foundation/src/index.ts
@@ -1,0 +1,1 @@
+export { genRandomChars } from './utils';
