+++ packages/foundation/src/utils.ts
@@ -1,0 +1,5 @@
+export const genRandomChars = (len = 6): string =>
+  Math.random()
+    .toString(36)
+    .replace(/[^a-z0-9]+/g, '')
+    .substr(1, len);
